+++ libft/ft_tabnew_two.c
@@ -6,7 +6,7 @@
 /*   By: cfatrane <cfatrane@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/01/21 14:22:42 by cfatrane          #+#    #+#             */
-/*   Updated: 2017/01/21 14:22:50 by cfatrane         ###   ########.fr       */
+/*   Updated: 2017/01/27 20:29:24 by cfatrane         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 int		**ft_tabnew_two(int x, int y)
 {
-	int i;
-	int **tab;
-	int *tab2;
+	size_t	i;
+	int		**tab;
+	int		*tab2;
 
 	i = 0;
 	if (!(tab = (int**)malloc(sizeof(tab) * y)))
